Prevent bug where event handler doesn't fire on first rigidbody created.
rapier.js library returns 0 as a handle for the first rigidbody when it's created in this line: https://github.com/pmndrs/react-three-rapier/blob/c9dd1a7b74e8fb7f07aa967f29630a78d3be58a1/packages/react-three-rapier/src/hooks.ts#L56
Because the value is 0, it fails the check changed here, which means the events on the first rigidbody don't fire. This can be tested by going to the demo scene Transforms and reloading the scene. Without this change, the console log in the first RigidBody will not fire. The first index being 0 is intended behaviour, according to the rapier.js maintainer.